ref(normalization): Restore SchemaProcessor
